Added photo upload and retrieval functionality to Profile

+++ Eventopia.Infra/Service/ProfileService.cs
@@ -1,7 +1,8 @@
 ﻿using Eventopia.Core.Repository;
 using Eventopia.Core.Service;
 using Eventopia.Core.Data;
-
+using Eventopia.Infra.Repository;
+using Eventopia.Infra.Utility;
 
 namespace Eventopia.Infra.Service;
 
@@ -16,12 +17,23 @@ public class ProfileService : IProfileService
 
     public Profile GetById(int id)
     {
-        return _profileRepository.GetById(id);
+		Profile profile = _profileRepository.GetById(id);
+		if (profile == null)
+			return null;
+		byte[]? byteFile = ImageUtility.RetrieveImage(profile.ImagePath, "Profile");
+		profile.RetrievedImageFile = byteFile;
+		return profile;
     }
 
     public List<Profile> GetAll()
     {
-        return _profileRepository.GetAll();
+		List<Profile> profiles = _profileRepository.GetAll();
+		foreach (Profile profile in profiles)
+		{
+			byte[]? byteFile = ImageUtility.RetrieveImage(profile.ImagePath, "Profile");
+			profile.RetrievedImageFile = byteFile;
+		}
+		return profiles;
     }
 
     public bool CreateNew(Profile profile)
@@ -36,11 +48,20 @@ public class ProfileService : IProfileService
 
     public bool Delete(int id)
     {
-        return _profileRepository.Delete(id);
+		Profile profile = _profileRepository.GetById(id);
+		if (profile == null)
+			return false;
+		ImageUtility.DeleteImage(profile.ImagePath, "Profile");
+		return _profileRepository.Delete(id);
     }
 
 	public Profile GetProfileByPhoneNumber(string phoneNumber)
 	{
-		return _profileRepository.GetProfileByPhoneNumber(phoneNumber);
+		Profile profile = _profileRepository.GetProfileByPhoneNumber(phoneNumber);
+		if (profile == null)
+			return null;
+		byte[]? byteFile = ImageUtility.RetrieveImage(profile.ImagePath, "Profile");
+		profile.RetrievedImageFile = byteFile;
+		return profile;
 	}
 }
